+++ wagtail_localize/__init__.py
@@ -1,10 +1,10 @@
-from .version import get_version
+from django import VERSION as DJANGO_VERSION
+
+from .version import VERSION, __version__  # noqa
 
 
-default_app_config = "wagtail_localize.apps.WagtailLocalizeAppConfig"
-
-
-# release must be one of alpha, beta, rc, or final
-VERSION = (1, 0, 1, "final", 1)
-
-__version__ = get_version(VERSION)
+if DJANGO_VERSION >= (3, 2):
+    # The declaration is only needed for older Django versions
+    pass
+else:
+    default_app_config = "wagtail_localize.apps.WagtailLocalizeAppConfig"