+++ wagtail_localize/models.py
@@ -36,6 +36,7 @@ from modelcluster.models import (
     get_serializable_data_for_fields,
     model_from_serializable_data,
 )
+from wagtail import VERSION as WAGTAIL_VERSION
 from wagtail.core import blocks
 from wagtail.core.fields import StreamField
 from wagtail.core.models import (
@@ -1280,6 +1281,7 @@ class Translation(models.Model):
                         string_translation.tool_name = tool_name
                         string_translation.last_translated_by = user
                         string_translation.updated_at = timezone.now()
+                        string_translation.has_error = False  # reset the error flag.
                         string_translation.save()
 
             except TranslationContext.DoesNotExist:
@@ -1513,6 +1515,10 @@ class TranslationContext(models.Model):
                             ] + get_field_path_from_stream_block(
                                 block.value, path_components[1:]
                             )
+                        elif isinstance(
+                            block_def, blocks.ListBlock
+                        ) and WAGTAIL_VERSION >= (2, 16):
+                            return [block.block_type, path_components[1]]
 
                         else:
                             return [block.block_type]