+++ wagtail_localize/tests/test_edit_translation.py
@@ -11,11 +11,17 @@ from django.contrib.auth.models import Group, Permission
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.messages import get_messages
 from django.core.files.uploadedfile import SimpleUploadedFile
-from django.test import TestCase
+from django.test import TestCase, override_settings
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import gettext_lazy
 from freezegun import freeze_time
+from rest_framework.authentication import SessionAuthentication, TokenAuthentication
+from rest_framework.permissions import (
+    DjangoModelPermissionsOrAnonReadOnly,
+    IsAuthenticated,
+)
+from rest_framework.settings import api_settings
 from rest_framework.test import APITestCase
 from wagtail.admin.edit_handlers import FieldPanel
 from wagtail.core.blocks import StreamValue
@@ -42,6 +48,10 @@ from wagtail_localize.test.models import (
     TestHomePage,
     TestPage,
     TestSnippet,
+)
+from wagtail_localize.views.edit_translation import (
+    edit_override,
+    edit_string_translation,
 )
 from wagtail_localize.wagtail_hooks import SNIPPET_RESTART_TRANSLATION_ENABLED
 
@@ -727,7 +737,7 @@ class TestGetEditTranslationView(EditTranslationTestData, TestCase):
             {
                 "id": str(url_block_id),
                 "type": "test_urlblock",
-                "value": "https://wagtail.io/",
+                "value": "https://wagtail.org/",
             },
             {
                 "id": str(page_block_id),
@@ -785,7 +795,7 @@ class TestGetEditTranslationView(EditTranslationTestData, TestCase):
                 (
                     f"test_streamfield.{url_block_id}",
                     {"type": "text"},
-                    "https://wagtail.io/",
+                    "https://wagtail.org/",
                 ),
                 (
                     f"test_streamfield.{page_block_id}",
@@ -2086,6 +2096,39 @@ class TestEditStringTranslationAPIView(EditTranslationTestData, APITestCase):
 
         self.assertEqual(response.status_code, 403)
 
+    @override_settings(
+        REST_FRAMEWORK={
+            "DEFAULT_PERMISSION_CLASSES": [
+                "rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly"
+            ]
+        }
+    )
+    def test_edit_translation_with_drf_default_permissions_classes(self):
+        self.assertEqual(
+            api_settings.DEFAULT_PERMISSION_CLASSES,
+            [DjangoModelPermissionsOrAnonReadOnly],
+        )
+        self.assertEqual(
+            edit_string_translation.view_class.permission_classes, [IsAuthenticated]
+        )
+
+    @override_settings(
+        REST_FRAMEWORK={
+            "DEFAULT_AUTHENTICATION_CLASSES": [
+                "rest_framework.authentication.TokenAuthentication"
+            ]
+        }
+    )
+    def test_update_override_with_drf_default_authentication_classes(self):
+        self.assertEqual(
+            api_settings.DEFAULT_AUTHENTICATION_CLASSES,
+            [TokenAuthentication],
+        )
+        self.assertEqual(
+            edit_string_translation.view_class.authentication_classes,
+            [SessionAuthentication],
+        )
+
 
 @freeze_time("2020-08-21")
 class TestEditOverrideAPIView(EditTranslationTestData, APITestCase):
@@ -2225,6 +2268,36 @@ class TestEditOverrideAPIView(EditTranslationTestData, APITestCase):
         )
 
         self.assertEqual(response.status_code, 403)
+
+    @override_settings(
+        REST_FRAMEWORK={
+            "DEFAULT_PERMISSION_CLASSES": [
+                "rest_framework.permissions.DjangoModelPermissionsOrAnonReadOnly"
+            ]
+        }
+    )
+    def test_update_override_with_drf_default_permissions_classes(self):
+        self.assertEqual(
+            api_settings.DEFAULT_PERMISSION_CLASSES,
+            [DjangoModelPermissionsOrAnonReadOnly],
+        )
+        self.assertEqual(edit_override.view_class.permission_classes, [IsAuthenticated])
+
+    @override_settings(
+        REST_FRAMEWORK={
+            "DEFAULT_AUTHENTICATION_CLASSES": [
+                "rest_framework.authentication.TokenAuthentication"
+            ]
+        }
+    )
+    def test_update_override_with_drf_default_authentication_classes(self):
+        self.assertEqual(
+            api_settings.DEFAULT_AUTHENTICATION_CLASSES,
+            [TokenAuthentication],
+        )
+        self.assertEqual(
+            edit_override.view_class.authentication_classes, [SessionAuthentication]
+        )
 
 
 class TestDownloadPOFileView(EditTranslationTestData, TestCase):